test: combine visualization.cy.js and table.cy.js into one file
